+++ tests/test_external_services/test_tprek/test_tprek_unit_hauki_resource_importer.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pytest
 
 from tilavarauspalvelu.integrations.opening_hours.hauki_api_client import HaukiAPIClient
-from tilavarauspalvelu.utils.tprek.tprek_unit_importer import TprekUnitHaukiResourceIdImporter
+from tilavarauspalvelu.integrations.tprek.tprek_unit_importer import TprekUnitHaukiResourceIdImporter
 
 from tests.factories import OriginHaukiResourceFactory, UnitFactory
 from tests.helpers import patch_method
