+++ tests/test_external_services/test_tprek/test_tprek_unit_importer.py
@@ -6,9 +6,9 @@ import re
 import pytest
 from django.utils.timezone import get_default_timezone
 
+from tilavarauspalvelu.integrations.tprek.tprek_api_client import TprekAPIClient
+from tilavarauspalvelu.integrations.tprek.tprek_unit_importer import TprekUnitHaukiResourceIdImporter, TprekUnitImporter
 from tilavarauspalvelu.models import Location
-from tilavarauspalvelu.utils.tprek.tprek_api_client import TprekAPIClient
-from tilavarauspalvelu.utils.tprek.tprek_unit_importer import TprekUnitHaukiResourceIdImporter, TprekUnitImporter
 from utils.date_utils import DEFAULT_TIMEZONE, local_datetime
 from utils.sentry import SentryLogger
