+++ tilavarauspalvelu/admin/unit/admin.py
@@ -10,8 +10,8 @@ from django.utils.translation import gettext_lazy as _
 from modeltranslation.admin import TranslationAdmin
 
 from tilavarauspalvelu.admin.location.admin import LocationInline
+from tilavarauspalvelu.integrations.tprek.tprek_unit_importer import TprekUnitImporter
 from tilavarauspalvelu.models import Unit
-from tilavarauspalvelu.utils.tprek.tprek_unit_importer import TprekUnitImporter
 from utils.sentry import SentryLogger
 
 if TYPE_CHECKING:
